refactored the API versions in better structure
